Remove hogan_assets gem and all related code
The gem hasn't been updated in years and is raising a lot of deprecation
warnings now. The maintainer of the gem doesn't seem to be too
interested in updating the gem so I replaced it with React, which is
something that I will use more in other parts of the app.